LayoutTests/webaudio/offlineaudiocontext-constructor.html

<!DOCTYPE html>
<html>
    <head>
        <title>offlineaudiodestinationnode-crash</title>
        <script src="../resources/js-test-pre.js"></script>
        <script>
            description('This tests that passing zero to the "numberOfFrames" parameter of the OfflineAudioContext constructor does not cause a crash.');
            shouldThrow('new webkitOfflineAudioContext(1, 0, 44100)');
            shouldThrow('new webkitOfflineAudioContext(0, 1, 44100)');
            shouldThrow('new webkitOfflineAudioContext(1, 1, 0)');
            shouldNotThrow('new webkitOfflineAudioContext(1, 1, 44100)');
            testPassed('did not crash');
        </script>
    </head>
    <body>
    </body>
</html>
